🎉 add : 스크롤 위치 추적 훅 및 인트로 애니메이션 상태 관리 추가 #3
