+++ src/hooks/useScrollPosition.ts
@@ -1,0 +1,47 @@
+import { ScrollData } from '@/components/background/types';
+import { useEffect, useState } from 'react';
+
+export const useScrollPosition = () => {
+  const [scrollData, setScrollData] = useState<ScrollData>({
+    scrollY: 0,
+    scrollVelocity: 0,
+    scrollProgress: 0,
+  });
+
+  useEffect(() => {
+    let lastScrollY = window.scrollY;
+    let ticking = false;
+
+    const updateScrollData = () => {
+      const currentScrollY = window.scrollY;
+      const scrollVelocity = currentScrollY - lastScrollY;
+      const scrollProgress =
+        currentScrollY /
+        (document.documentElement.scrollHeight - window.innerHeight);
+
+      setScrollData({
+        scrollY: currentScrollY,
+        scrollVelocity,
+        scrollProgress,
+      });
+
+      lastScrollY = currentScrollY;
+      ticking = false;
+    };
+
+    const onScroll = () => {
+      if (!ticking) {
+        window.requestAnimationFrame(updateScrollData);
+        ticking = true;
+      }
+    };
+
+    window.addEventListener('scroll', onScroll, { passive: true });
+
+    return () => {
+      window.removeEventListener('scroll', onScroll);
+    };
+  }, []);
+
+  return scrollData;
+};
